move render input field in common folder
